compile mseed with PIC; rename the new module/namespace as io
