+++ cxx/include/mspass/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory(seismic)
 add_subdirectory(utility)
 add_subdirectory(algorithms)
-add_subdirectory(import)
+add_subdirectory(io)
