--- cxx/include/mspass/import/CMakeLists.txt
@@ -1,2 +1,0 @@
-FILE(GLOB import_headers *.h)
-install (FILES ${import_headers} DESTINATION include/mspass/import)
