+++ cxx/include/mspass/io/CMakeLists.txt
@@ -1,0 +1,2 @@
+FILE(GLOB io_headers *.h)
+install (FILES ${io_headers} DESTINATION include/mspass/io)
