+++ cxx/python/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory (algorithms)
 add_subdirectory (seismic)
 add_subdirectory (utility)
-add_subdirectory (importers)
+add_subdirectory (io)
