+++ cxx/python/io/CMakeLists.txt
@@ -9,10 +9,10 @@ include_directories(
 #add_library(mspasspython STATIC ${sources_python})
 #target_link_libraries(mspasspython PRIVATE mspass ${PYTHON_LIBRARIES})
 
-pybind11_add_module(pyimport import_py.cc)
+pybind11_add_module(pyio io_py.cc)
 
-target_link_libraries(pyimport PRIVATE import ${MSEED_LIBRARIES})
-set_target_properties(pyimport PROPERTIES OUTPUT_NAME "importers")
+target_link_libraries(pyio PRIVATE mspass ${MSEED_LIBRARIES})
+set_target_properties(pyio PROPERTIES OUTPUT_NAME "io")
 if(DEFINED CMAKE_LIBRARY_OUTPUT_DIRECTORY)
-  set_target_properties(pyimport PROPERTIES LIBRARY_OUTPUT_DIRECTORY ${CMAKE_LIBRARY_OUTPUT_DIRECTORY}/ccore)
+  set_target_properties(pyio PROPERTIES LIBRARY_OUTPUT_DIRECTORY ${CMAKE_LIBRARY_OUTPUT_DIRECTORY}/ccore)
 endif()
