+++ cxx/src/lib/CMakeLists.txt
@@ -1,9 +1,9 @@
 add_subdirectory(utility)
 add_subdirectory(seismic)
 add_subdirectory(algorithms)
-add_subdirectory(import)
+add_subdirectory(io)
 
-add_library(mspass $<TARGET_OBJECTS:seismic> $<TARGET_OBJECTS:utility> $<TARGET_OBJECTS:alg_basics> $<TARGET_OBJECTS:amplitudes> $<TARGET_OBJECTS:deconvolution>)
-target_link_libraries(mspass PRIVATE ${BLAS_LIBRARIES} ${LAPACK_LIBRARIES} ${YAML_CPP_LIBRARIES} ${PYTHON_LIBRARIES} ${GSL_LIBRARIES})
+add_library(mspass $<TARGET_OBJECTS:seismic> $<TARGET_OBJECTS:utility> $<TARGET_OBJECTS:alg_basics> $<TARGET_OBJECTS:amplitudes> $<TARGET_OBJECTS:deconvolution> $<TARGET_OBJECTS:io>)
+target_link_libraries(mspass PRIVATE ${BLAS_LIBRARIES} ${LAPACK_LIBRARIES} ${YAML_CPP_LIBRARIES} ${PYTHON_LIBRARIES} ${GSL_LIBRARIES} ${MSEED_LIBRARIES})
 
 install (TARGETS mspass DESTINATION lib)
