+++ cxx/src/lib/io/CMakeLists.txt
@@ -1,0 +1,8 @@
+FILE(GLOB sources_io *.cc)
+include_directories(
+  ${MSEED_INCLUDE_DIR}
+  ${PROJECT_BINARY_DIR}/include
+  ${PROJECT_SOURCE_DIR}/include
+)
+
+add_library(io OBJECT ${sources_io})
